Add "join lobby"-button
The browser extension will insert a join lobby button on the
wikiweaverserver domain specified by the user. When pressed, this will
join the currently open lobby with stored username. The user most have
specified a username beforehand.

Note: it uses the activeTab permission (instead of permission on
allurls) to be able to insert itself on only the currently active tab.
Thismeans that the user must be on the website and open (or have open)
theextension. The current tab is only part of the activeTab
permissionbased on user interaction, e.g. opening the extension.

https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/manifest.json/permissions#activetab_permission
Signed-off-by: Victor Löfgren <victor.lofgren99@pm.me>